Introduce ReflectionAbstract.
Reduces duplicated code in BaseReflector and FileReflector.
