+++ src/phpDocumentor/Application.php
@@ -138,11 +138,10 @@ class Application extends \Cilex\Application
      */
     protected function linkEventDispatcherToSuperclasses()
     {
-        \phpDocumentor\Reflection\BaseReflector::$event_dispatcher
-            = \phpDocumentor\Reflection\FileReflector::$event_dispatcher
-                = \phpDocumentor\Parser\ParserAbstract::$event_dispatcher
-                    = \phpDocumentor\Transformer\TransformerAbstract::$event_dispatcher
-                        = $this['event_dispatcher'];
+        \phpDocumentor\Reflection\ReflectionAbstract::$event_dispatcher
+            = \phpDocumentor\Parser\ParserAbstract::$event_dispatcher
+                = \phpDocumentor\Transformer\TransformerAbstract::$event_dispatcher
+                    = $this['event_dispatcher'];
     }
 
     protected function loadPlugins()
