+++ src/phpDocumentor/Reflection/FileReflector.php
@@ -19,7 +19,7 @@ namespace phpDocumentor\Reflection;
  * @license  http://www.opensource.org/licenses/mit-license.php MIT
  * @link     http://phpdoc.org
  */
-class FileReflector extends \PHPParser_NodeVisitorAbstract
+class FileReflector extends ReflectionAbstract implements \PHPParser_NodeVisitor
 {
     protected $hash = array('a' => 1, 'b' => 2);
     protected $contents = 1;
@@ -44,13 +44,6 @@ class FileReflector extends \PHPParser_NodeVisitorAbstract
 
     protected $namespace_aliases = array();
     protected $current_namespace = '';
-
-    /**
-     * The event dispatcher object, may be null to not dispatch events.
-     *
-     * @var \sfEventDispatcher|null
-     */
-    public static $event_dispatcher = null;
 
     /**
      * Opens the file and retrieves its contents.
@@ -400,40 +393,6 @@ class FileReflector extends \PHPParser_NodeVisitorAbstract
         $this->filename = $filename;
     }
 
-    /**
-     * Dispatches an event to the Event Dispatcher.
-     *
-     * This method tries to dispatch an event; if no Event Dispatcher has been
-     * set than this method will explicitly not fail and return null.
-     *
-     * By not failing we make the Event Dispatcher optional and is it easier
-     * for people to re-use this component in their own application.
-     *
-     * @param string   $name      Name of the event to dispatch.
-     * @param string[] $arguments Arguments for this event.
-     *
-     * @throws \phpDocumentor\Reflection\Exception if there is a dispatcher but
-     *  it is not of type sfEventDispatcher
-     *
-     * @return mixed|null
-     */
-    public function dispatch($name, $arguments)
-    {
-        if (!self::$event_dispatcher) {
-            return null;
-        }
-
-        if (!self::$event_dispatcher instanceof \sfEventDispatcher) {
-            throw new \phpDocumentor\Reflection\Exception(
-                'Expected the event dispatcher to be an instance of '
-                . 'sfEventDispatcher'
-            );
-        }
-
-        $event = self::$event_dispatcher->notify(
-            new \sfEvent($this, $name, $arguments)
-        );
-
-        return $event ? $event->getReturnValue() : null;
-    }
+    public function leaveNode(\PHPParser_Node $node) { }
+    public function afterTraverse(array $nodes)      { }
 }
